Rearrange code related to holding locations
* Move the ruby-specific code to ruby_bindings module
* Move the holding location-specific code to the holding_locations module
